Rework sandctl utility
Original sandctl utility is inconvenient to use, particularly
if the guest code is quite large. Therefore, it was decided
to add the ability to load guest code from file. The file
must be in BFD "binary" format. The new mode is activated
with "-f" flag. However, the legacy mode is retained for
compatibility reasons.

Bug fixes:
-- #4: Possibility of code buffer overflow.
-- #5: Closing non-valid file descriptor (EBADF).